coherent results using fuzzy tokens

+++ src/main/java/matching/computers/names/NameDistanceComputer.java
@@ -2,5 +2,5 @@ package matching.computers.names;
 
 public interface NameDistanceComputer {
 
-    public double computeDistanceBetweenTwoNames(String name1, String name2);
+    public int computeDistanceBetweenTwoNames(String name1, String name2);
 }
